Refactor: move where args are converted to their schema form.
Previously, we did it in `GraphQLResolver#call`, which gets invoked as
part of resolving _every_ GraphQL field at every level of the response
structure. We don't want to convert args to their schema form if we're
not going to do anything with the args so it's better to defer it until
it's needed.