Refactor: move where we call `field.coerce_result`.
Previously, we called it from `GarphQLAdapter#call`, applying
it to all resolved fields. However, we only need to use it in the
aggregation `GroupedBy` resolver, where we get `DayOfWeek` enum
values that we may need to coerce. There are no other cases where
it is needed.

This allows us to simplify `GarphQLAdapter#call` so that it now only
does two things:

* Identifies the resolver to dispatch to.
* Dispatches to that resolver.

This will allow us to further optimize by leveraging built-in
functionality of the GraphQL gem to have it dispatch to the appropriate
resolver in a more performant manner.