Refactor: split `GetRecordFieldValue` resolver.
Previously, it handled both the `Hash` and `Document` cases, via a `case`
statement on `object`. Instead, we can handle it with two separate
resolvers. The `Document` resolver simply delegates to the `Hash`
resolver after extracting the hash `payload`.

We want to minimize the logic that needs to happen to resolve each
field, so it's nice to eliminate this `case` statement.