Rework how `Query` type is defined and customized.
Previously, it was defined in `GraphQLSDLEnumerator` and required an extension
to customize it. Defining it in `GraphQLSDLEnumerator` meant that it was only
included in the `schema.graphql` artifact and was not available to dump in
`runtime_metadata.yaml`. We need to be able to dump `resolver` runtime metadata
for the `Query` type, and this enables that. Customization can be done via
`on_built_in_types`.