Update resolver interface: pass a schema field.
Previously, we passed separate `parent_type` and `graphql_field`
arguments. This was done to align our resolver `#call` interface
with the resolver interface of the GraphQL gem. However, as per
the docs[^1], when we switch to providing an an implementation
hash for `default_resolve`, the interface is different--each resolver
is a lambda that accepts 3 arguments `(obj, args, ctx)`. Our resolvers
need to know what the field is, an we'll need to adapt the 3-arg
interface to our own resolver interface. As part of that, we'll provide
an `ElasticGraph::GraphQL::Schema::Field` object to each resolver
as the first argument.

[^1] https://graphql-ruby.org/schema/sdl.html