Update project template factories so they don't produce as much data.
The original factories produced such a volume of data that a simple
query to fetch every field returned ~5 MB of data. That wound up being
a slow query (given how much data was being returned), and took longer
to bootstrap.

In addition, I've updated the artist and venue factories to avoid producing
multiple records with the same `name` as that is a bit confusing.