Test the website example queries from `config/docker_demo/test`.
Now that the website encourages users to boot our docker demo
from the pages with example queries, I'd like to make sure the
example queries return actual data. To enforce that, I've updated
the CI build so that when the docker image is booted, we test each
of our queries, validating that it the response was non-empty.

As it turns out, 12 of our 41 example queries do not return data.
For now, I've added these 12 to an `allowed_failures` set. In future
PRs, I'll work through these and fix them.