Move API docs from `/docs` to `/api-docs`.
Our site features documentation beyond the YARD API docs, and it's
clearer to use `/api-docs` as the path for the API docs. It also
aligns with the way we label it at the website.

Note: ideally, we'd redirect from the `/docs` URLs to the new `/api-docs`
URLs, but there's not an easy way to do that with Jekyll, and if we're
going to break some links, now is the time to do it.