Return a clear error when indices have not been configured.
Previously, we'd silently return no results (which was ok) in most
cases, but would return a 500 internal server error when executing
a sub-aggregations query with aggregated values. We don't get back
the sort of response the sub aggregations logic expects when there
are no indices.

Now we return a clear error indicating the indices should be configured.