Remove `for_context` argument from `SchemaArtifacts::FromDisk`.
Previously, it was used to deal with `graphql_extension_modules` in
the runtime metadata. In non-graphql contexts, we could not count on
these extension modules being avilable to load. The `for_context`
argument indicated whether or not we were in a GraphQL context and
should load them or not.

However, this setup was suboptimal: it "infected" lots of places with
the extra `for_context` argument (e.g. each callsite, and each of their
callsites, etc...) and it makes things harder to reason about when the
runtime metadata loads differently for one context vs another.

Instead, I've added a layer of indirection (a new `GraphQLExtension`
class) which holds a reference to an unloaded extension, and allows the
extension to be loaded on demand. That allows `elasticgraph-graphql`
to load the extension when needed, while avoiding loading it in other
contexts, without the need for the `for_context` argument. It also
aligns with how we deal with our other extension modules in runtime
metadata -- they are also loaded lazily on demand rather than eagerly
when loading `runtime_metadata.yaml`.