Improve `script/update_gem_constraints`.
In #501, dependabot updated the `elasicsearch` version constraint from this:

```ruby
"~> 8.17", ">= 8.17.2"
```

To this:

```ruby
">= 8.17.2", "< 10.0"
```

When GitHub actions then ran `script/update_gem_constraints`, it then updated
the version constraint to:

```ruby
"~> 9.0.0"
```

However, this is tighter than we want it to be. Our original constraint
allowed any `8.x` version so long as it was `8.17.2` or greater by using
two digits with the `~>` operator. We want to maintain the same number of
digits (`"~> 9.0"`).

Here I used goose to update the script so that it does that. It uses the
version constraints from the commit before `HEAD` to determine the level
of precision we were using before dependabot updated the constraint, and
then carries that forward.