Add support for `IGNORE_SQS_LATENCY_TIMESTAMPS_FROM_ARNS`.
This can be used from `elasticgraph-indexer_lambda` to ignore
SQS latency timestamps for messages coming from a specific SQS
queue. This is useful, for example, when dealing with an indexer
lambda which ingests messages from multiple queues which are
treated as being at different priorities. A low-pri queue that is
intentionally throttled could skew overall latency metrics and
it's nice to be able to ignore the latency timestamps for messages
from such a queue.