Provide fully bundled GraphiQL UI from a new elasticgraph-graphiql gem.
Previously, we served an old version of GraphiQL using a simple `index.html`
page which relied on a CDN. This had a couple problems:

* If the CDN had problems, it would stop working.
* You had to be online to use it.

I've replaced that with a new approach that bundles all assets. It does
so using the graphiql-vite example from here:

https://github.com/graphql/graphiql/tree/graphiql%405.0.4/examples/graphiql-vite

I've included a script (written by goose) which bundles the assets. The assets are
bundled into a tarball because there are so many files and they aren't something
we want to individually review in PRs.

Unfortunately, the tarball is over 2 MB, which I felt was too large to include
in `elasticgraph-rack` (which is less than 10 KB!). So, I've extracted the GraphiQL
UI into `elasticgraph-graphiql` and projects which do not need a GraphiQL UI
don't need to pull in such a large dependency.