Upgrade GraphiQL to v5.0.0 and bundle all assets.
Previously, we served an old version of GraphiQL which relied on a CDN.
If the CDN had problems, it would stop working, and you had to be online
to use it.

I've replaced that with a new approach that bundles all assets. It does
so using the graphiql-vite example from here:

https://github.com/graphql/graphiql/tree/graphiql%405.0.0/examples/graphiql-vite

I've included a script (written by goose) which bundles the assets. The
assets are bundled into a tarball because there are so many files and they
aren't something we want to individually review in PRs.