chore: upgrade bundler to 2.6.9.
The new version of bundler includes some internal refactorings that
broke our monorepo setup. To stay compatible, I've had to rework our
`gem` override to instead override `add_dependency`.

I believe this is the root of the problem I reported to dependabot:

https://github.com/dependabot/dependabot-core/issues/12426

The latest dependabot uses Bundler 2.6.9, and this was causing the
issues reported in that ticket since our `gem` override was no longer
effective.

The specific bundler change prompting this is this PR:

https://github.com/rubygems/rubygems/pull/8480

In that PR, the `gemspec` method was refactored to no longer call `gem`.
Instead, it calls a new `add_dependency` method.