fix: fail with a clear error when `nested` mapping type is improperly used.
The `nested` mapping type is only valid on a list-of-objects field
(or a `paginated_collection_field` with an element type that is an object).

If `nested` is used on a list-of-scalars field, confusing GraphQL schema
errors could occur, such as:

```
GraphQL::Schema::InvalidDocumentError: Type "TeamStringSubAggregationConnection" not found in document.
```

This avoids those confusing errors.