Merge `elasticgraph-json_schema` into `elasticgraph-support`.
I'm working on integrating JSON schema validation into the ElasticGraph
config system. As part of that, I'm defining a base config class that
handles JSON schema validation. The natural place to define that is in
`elasticgraph-support` but our JSON schema support needs to be in there
before we can do that.

Previously, we kept `elasticgraph-json_schema` separate since JSON
schema wasn't used in all execution contexts (such as GraphQL), and
we didn't want to cause the `json_schemer` dependency to get pulled
into contexts where it wasn't used. That'll no longer be a concern
with the new usage of JSON schema in our config system as all parts
of ElasticGraph will use the config system.