Introduce `ElasticGraph::Config`.
This defines a configuration class which uses JSON schema to validate
the provided configuration, removing the need to perform ad-hoc config
validation like we were before. In addition, the JSON schema will allow
us to improve our documentation as we can merge the JSON schemas from
all configuration classes to provide the overall config schema of
ElasticGraph in our docs. (That will come in a later PR).

For now, this PR:

- Defines `ElasticGraph::Config` and tests it.
- Defines shared specs in `elasticgraph/spec/unit/elastic_graph/gem_spec.rb` which
  extensively tests the config classes.
  - It verifies that all config classes are defined using `ElasticGraph::Config`
  - It verifies that each config class has a valid JSON schema.
  - It verifies that each config class provides a `default` value for any properties which are not `required`.
  - It verifies that all `default` and `examples` values are valid.
  - It verifies that every property has a `description` and `examples` (which I plan to use in our docs later).
  - It verifies that each config class has a unique `path`.
- Uses the new config system for one class (`ElasticGraph::Support::Logger::Config`) and excludes
  the others from being tested in `gem_spec.rb` (they'll be converted and tested in a later PR.)