site: widen articles a bit, so wider content can fit in code boxes.
This adds ~20 characters of width to the center `article` element
which wraps the content of most pages.

The direct impetus for this is a new configuration guide I'm working
on. I am pulling in a generated example config and JSON schema config
and they are wider than the article width. As a result, the code boxes
have to include horizontal scrolling. While horizontal scrolling isn't
the worst to include, I think our `article` elements can be a bit wider.
This looks a bit better to me overall and will allow wider code snippets
to render without horizontal scrolling.