Improve `script/update_gem_constraints`: support pinned versions.
The prior version of this script assumed no dependencies were pinnned
to a specific version. This meant that it would update our currently
pinned version of the `graphql` gem (`"2.5.11"`), changing it to
`"~> 2.5.11"` which allows 2.5.12 and later to be used, which we want
to avoid.

Now the script respects pinned versions while still properly updating
`>=` and `~>` dependency requirements.

This should allow us to safely merge some of the dependabot PRs we've
held off on.