Upgrade GraphQL gem from 2.5.11 to 2.5.14. (#919)
We had been pinned to 2.5.11 due to an `elasticgraph-apollo` failure,
but since then we've pulled in an `apollo-federation` upgrade which
fixes the issue.

In addition, 2.5.13 of the GraphQL gem has a regression which impacts
a couple of doc tests:

https://github.com/rmosolgo/graphql-ruby/issues/5466

However, our doc test was not actually correct--it should be passing
arguments as a list-of-lists. Once we pass the argument correctly, we
can safely upgrade the GraphQL gem to 2.5.14.